fix bug in loading levels add som soundes. super_pacman's sound not work well

+++ src/Timer.cpp
@@ -25,10 +25,10 @@ void Timer::setStart(size_t board_num)
 {
   switch (board_num)
   {
-//  case 1:  m_time_start = 10;  break;
-  case 1:  m_time_start = static_cast<size_t>(3) * 60 + 1;  break;
-  case 2:  m_time_start = static_cast<size_t>(4) * 60 + 1;  break;
-  case 3:  m_time_start = static_cast<size_t>(5) * 60 + 1;  break;
+  //case 1:  m_time_start = 10;  break;
+  case 1:  m_time_start = 3 * 60 + 1;  break;
+  case 2:  m_time_start = 4 * 60 + 1;  break;
+  case 3:  m_time_start = 5 * 60 + 1;  break;
   }
   m_elapsed = 0;
 }
@@ -44,12 +44,6 @@ std::string Timer::getTimer() const
 {
   return m_timer;
 }
-
-//-------------------------------------------------------------------
-//void Timer::pause()
-//{
-//
-//}
 
 //-------------------------------------------------------------------
 sf::Color Timer::getColor() const
